requirements.txt
psycopg2==2.8.1

pytest==4.0.1
pytest-faker==2.0.0
pytest-repeat==0.8.0
pytest-benchmark==3.2.2

flake8==3.5.0
flake8-isort==2.5
flake8-commas==2.0.0
flake8-print==3.1.0
isort==4.3.4
